% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_api_version.R
\name{grib_api_version}
\alias{grib_api_version}
\title{Print ecCodes version}
\usage{
grib_api_version()
}
\value{
Returns a string of the format MAJOR.MINOR.REVISION.
}
\description{
\code{grib_api_version} prints the ecCodes library version.
}
\examples{
grib_api_version()
}
