% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_cube.R
\name{grib_cube}
\alias{grib_cube}
\title{Create 3D volume of a GRIB variable}
\usage{
grib_cube(gribObj, shortName, typeOfLevel, decreasing = FALSE)
}
\arguments{
\item{gribObj}{\code{GRIB} class object.}

\item{shortName}{The short name given in the GRIB file of the variable to
select.}

\item{typeOfLevel}{The vertical coordinate to use as given by the typeOfLevel
key in the GRIB file.}

\item{decreasing}{Parameter to tell the array's vertical coordinate to be
increasing or decreasing.}
}
\value{
Returns a three-dimenional array.
}
\description{
\code{grib_cube} creates a three-dimensional array from one variable along a
chosen vertical coordinate.
}
\details{
\code{grib_cube} is a wrapper function for \code{grib_select} to conveniently
create a three-dimensional cube. The user inputs a variable to search for and
the vertical coordinate to use when finding each level.

Because \code{grib_cube} uses \code{grib_select}, speed can become an issue.
This is meant as a convenience to "get the job done". If you want more speed,
it will always be better to know which message number you want, set up your
own loop, and use \code{grib_get_message} as that will avoid the overhead of
searching through the GRIB file.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
cube <- grib_cube(g, 'u', 'isobaricInhPa', TRUE)
grib_close(g)
}
\seealso{
\code{\link{grib_get_message}} \code{\link{grib_list}}
  \code{\link{grib_expand_grids}} \code{\link{grib_latlons}}
  \code{\link{grib_select}}
}
