% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_get_message.R
\name{grib_get_message}
\alias{grib_get_message}
\title{Get GRIB messages}
\usage{
grib_get_message(gribObj, messages, mask = FALSE)
}
\arguments{
\item{gribObj}{GRIB class object.}

\item{messages}{an integer or vector of integers corresponding to the
messages to extract from the GRIB file.}

\item{mask}{optional logical that controls whether or not missing values are
changed to \code{NA} based on the bitmap and GRIB files missing value.}
}
\value{
Returns a \code{gribMessage} object.
}
\description{
\code{grib_get_message} retrieves a GRIB message from a GRIB file using the
message number
}
\details{
\code{grib_get_message} is a quick and easy way to retrieve a GRIB message
from a GRIB file if you know what message number it is. An easy way to
determine the message number is to use \code{\link{grib_list}} first.
}
\seealso{
\code{\link{grib_select}} \code{\link{grib_list}}
}

