% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_length.R
\name{grib_length}
\alias{grib_length}
\title{Get number of GRIB messages}
\usage{
grib_length(gribObj)
}
\arguments{
\item{gribObj}{\code{GRIB} class object.}
}
\value{
Returns \code{integer} number of GRIB messages.
}
\description{
\code{grib_length} loops through a GRIB file to determine the number of
messages.
}
\details{
\code{grib_length} determines the length of both GRIB files with and without
multiple fields per message. The GRIB file is looped though twice, with and
without multi-message support turned on, to get the true length.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
grib_length(g)
grib_close(g)
}
