% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_open.R
\name{grib_open}
\alias{grib_open}
\title{Open GRIB files}
\usage{
grib_open(file)
}
\arguments{
\item{file}{\code{character} string. A path to a GRIB file.}
}
\value{
Returns a \code{GRIB} class object: see 'Details'.
}
\description{
\code{grib_open} opens a GRIB file for reading or writing and returns a GRIB
class object.
}
\details{
\code{grib_open} is the constructor function for the S3 GRIB class object.
The GRIB object is list with three elements: \code{file}, \code{handle}, and
\code{isMultiMessage}. \code{file} is the path to the opened file,
\code{handle} is an \code{externalptr} to the \code{file} in memory, and
\code{isMultiMessage} is a \code{logical} indicating if there are multiple
entries in any of the \code{file}'s messages.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
g # shows status
grib_close(g)
}
\seealso{
\code{\link{grib_close}}
}
