% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_proj4str.R
\name{grib_proj4str}
\alias{grib_proj4str}
\title{Get PROJ4 string for a GRIB file}
\usage{
grib_proj4str(gribMessage, userProj4 = NULL)
}
\arguments{
\item{gribMessage}{gribMessage class object.}

\item{userProj4}{An optional argument that can be a named list or named
character vector of additional PROJ4 elements the user would like to add.}
}
\value{
Returns a character string of the PROJ4 elements associate with the
  \code{gribMessage}.
}
\description{
\code{grib_proj4str} generates a PROJ4 string for a given gribMessage class
object based on the message's keys and, optionally, user input
}
\details{
\code{grib_proj4str} uses a GRIB message's keys to generate a PROJ4 string on
the fly. A user can even input their own PROJ4 elements using the
\code{userProj4} arugment. Several other packages have functions that can use
this PROJ4 string to appropriately project the values in the GRIB message.
Several common grid types are supported at this time, but likely not all.
This function is heavily influenced the the PROJ4 string method used in the
pygrib Python module (see
\href{https://github.com/jswhit/pygrib/blob/master/pygrib.pyx}{pygrib source}).
}

