% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_select.R
\name{grib_select}
\alias{grib_select}
\title{Select multiple GRIB messages}
\usage{
grib_select(gribObj, keyPairs)
}
\arguments{
\item{gribObj}{\code{GRIB} class object.}

\item{keyPairs}{a named \code{list} of key/value pairs that will be used to
search through and match messages in the GRIB file using an index.}
}
\value{
Returns a \code{gribMessage} or a list of \code{gribMessage} objects
  if more than one message is selected.
}
\description{
\code{grib_select} selects all GRIB messages that match key/value pairs input
by the user.
}
\details{
\code{grib_select} is a function more robust than \code{grib_get_message} in
that it can search through a GRIB file and get all matching messages given
key(s)/value(s) pairs. This is achieved by making an index of the input GRIB
file and then selecting keys that match user input in \code{keyPairs}. The
values returned in the GRIB message are masked if they are coded as a missing
value or the bitmap, if present, masks them.

All grid (latitude, longitude, values, etc.) are output as vectors for
simplicity and flexibility. Many plotting functions want input in
\code{matrix} form (e.g., \code{image()}, \code{fields::image.plot()} while
others want input in long form \code{data.frame}s or vectors (e.g.,
\code{ggplot}). Keeping vector form allows the most flexibility for the user
to do what is needed to format the data appropriately. Some helper functions
in this package will do some data formatting for you.

\code{grib_select} is meant as an easy way to grab the data you want.
However, it can suffer from performance issues due to the overhead of
searching through the GRIB file. If speed is important, knowing the message
number in the file and using \code{grib_get_message} is the way to go.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
# select all mixing ratio messages
gm1 <- grib_select(g, list(shortName = "q"))
# select mixing ratio, but only at 1000 and 850 hPa
gm2 <- grib_select(g, list(shortName = "q", level = c(1000, 850)))
grib_close(g)
}
\seealso{
\code{\link{grib_get_message}} \code{\link{grib_list}}
  \code{\link{grib_expand_grids}} \code{\link{grib_latlons}}
}
