% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_key.R
\name{has.key}
\alias{has.key}
\title{Test for keys}
\usage{
has.key(gribMessage, key)
}
\arguments{
\item{gribMessage}{gribMessage object to be tested}

\item{key}{character string of the key to be tested for}
}
\value{
Returns a logical
}
\description{
\code{has.key} tests whether a gribMessage object has a given key
}

