% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grib_message.R
\name{is.grib.message}
\alias{is.grib.message}
\title{GRIB Messages}
\usage{
is.grib.message(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
Returns a logical
}
\description{
\code{is.grib.message} tests whether and object is a GRIB message
}
\seealso{
\code{\link{is.grib}}
}

