% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grib.R
\name{is.grib}
\alias{is.grib}
\title{GRIB Messages}
\usage{
is.grib(obj)
}
\arguments{
\item{obj}{object to be tested.}
}
\value{
Returns a \code{logical}.
}
\description{
\code{is.grib} tests whether and object is a GRIB object.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
gm <- grib_get_message(g, 1)
is.grib.message(g)
# should be FALSE
is.grib(gm)
grib_close(g)
}
\seealso{
\code{\link{is.grib.message}}
}
