% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.grib}
\alias{print.grib}
\title{Print GRIB Details}
\usage{
\method{print}{grib}(x, ...)
}
\arguments{
\item{x}{\code{GRIB} class object.}

\item{...}{futher arguments; ignored here.}
}
\description{
\code{print.grib} displays basic status information about a GRIB file.
}
\details{
\code{print.grib} is the S3 method for printing GRIB files. Its behavior
differs from most other print methods in that it displays some basic
information about the GRIB file. When invoked on a GRIB class object, it will
display the location of the GRIB file, whether the GRIB file has an active
file handle in memory, and whether or not the file containins multiple fields
in some messages.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
print(g)
grib_close(g)
}
