% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_select.R
\name{grib_select}
\alias{grib_select}
\title{Select multiple GRIB messages}
\usage{
grib_select(gribObj, keyPairs, mask = FALSE)
}
\arguments{
\item{gribObj}{GRIB class object.}

\item{keyPairs}{a named list of key/value pairs that will be used to search
through and match messages in the GRIB file using an index}

\item{mask}{optional logical that controls whether or not missing values are
changed to \code{NA} based on the bitmap and GRIB files missing value.}
}
\value{
Returns a list of gribMessage objects
}
\description{
\code{grib_select} selects all GRIB messages that match key/value pairs input
by the user.
}
\details{
\code{grib_select} is a function more robust than \code{grib_get_message} in
that it can search through a GRIB file and get all matching messages given
key(s)/value(s) pairs. This is achieved by making an index of the input GRIB
file and then selecting keys that match user input in \code{keyPairs}.

The \code{filter} parameter controls which keys will be filtered out of the
results. More information can be found in the
\href{https://software.ecmwf.int/wiki/display/GRIB/grib_api.h+File+Reference}{
GRIB API Reference}. Below are the options used in this package:
}
\seealso{
\code{\link{grib_get_message}} \code{\link{grib_list}}
}

