% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grib_select.R
\name{grib_select}
\alias{grib_select}
\title{Select multiple GRIB messages}
\usage{
grib_select(gribObj, keyPairs, mask = FALSE, filter = "none",
  nameSpace = "")
}
\arguments{
\item{gribObj}{GRIB class object.}

\item{keyPairs}{a named list of key/value pairs that will be used to search
through and match messages in the GRIB file using an index}

\item{mask}{optional logical that controls whether or not missing values are
changed to \code{NA} based on the bitmap and GRIB files missing value.}

\item{filter}{optional character string that controls what keys will be
filtered out of the resulting GRIB message. The default is "none". See
'Details' for other options}

\item{nameSpace}{optional character string that can control what special,
pre-defined group of keys gets return. Defaults to returning all. See
'Details' for full description.}
}
\value{
Returns a list of gribMessage objects
}
\description{
\code{grib_select} selects all GRIB messages that match key/value pairs input
by the user.
}
\details{
\code{grib_select} is a function more robust than \code{grib_get_message} in
that it can search through a GRIB file and get all matching messages given
key(s)/value(s) pairs. This is achieved by making an index of the input GRIB
file and then selecting keys that match user input in \code{keyPairs}.

The \code{filter} parameter controls which keys will be filtered out of the
results. More information can be found in the
\href{https://software.ecmwf.int/wiki/display/GRIB/grib_api.h+File+Reference}{
GRIB API Reference}. Below are the options used in this package:

\itemize{
  \item "none": no keys filtered
  \item "readonly": skip readonly keys
  \item "optional": skip optional keys
  \item "edition": skip edition-specific keys
  \item "coded": skip coded keys
  \item "computed": skip keys computed from other keys
  \item "duplicate: skip duplicate keys
  \item "func": skip keys that are functions
}

The \code{nameSpace} parameter is a quick and easy way to grab a subset of the
keys available in a GRIB file using pre-defined groups of related keys. More
information about GRIB key namespaces can be found
\href{https://software.ecmwf.int/wiki/display/GRIB/GRIB+API+Frequently+Asked
+Questions+FAQ#GRIBAPIFrequentlyAskedQuestionsFAQ-Whatarenamespaces?}{
here}. The \code{nameSpace} options used in this package are described
below:

\itemize{
  \item "" or \code{NULL}: all keys
  \item "ls": most commonly used keys
  \item "parameter": keys related to the parameter
  \item "statistics": keys related to statiscal values of the data
  \item "time": keys related to the forecast time/initialization
  \item "geography": keys describing the grid geometry
  \item "vertical": keys describing levels and layers
  \item "mars: ECMWF's Meteorological Archive and Retrieval System keys
}
}
\seealso{
\code{\link{grib_get_message}} \code{\link{grib_list}}
}

