% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_key.R
\name{has.key}
\alias{has.key}
\title{Test for keys}
\usage{
has.key(gribMessage, key)
}
\arguments{
\item{gribMessage}{gribMessage object to be tested.}

\item{key}{character string of the key to be tested for.}
}
\value{
Returns a \code{logical}
}
\description{
\code{has.key} tests whether a gribMessage object has a given key.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
gm <- grib_get_message(g, 1)
# this key exists
has.key(gm, "shortName")
# this key does not exist
has.key(gm, "grib")
grib_close(g)
}
