% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grib_message.R
\name{is.grib.message}
\alias{is.grib.message}
\title{GRIB Messages}
\usage{
is.grib.message(obj)
}
\arguments{
\item{obj}{object to be tested.}
}
\value{
Returns a \code{logical}.
}
\description{
\code{is.grib.message} tests whether and object is a GRIB message.
}
\examples{
g <- grib_open(system.file("extdata", "lfpw.grib1", package = "gribr"))
gm <- grib_get_message(g, 1)
is.grib.message(gm)
# should be FALSE
is.grib.message(g)
grib_close(g)
}
\seealso{
\code{\link{is.grib}}
}
