% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grib.R
\name{is.grib}
\alias{is.grib}
\title{GRIB Messages}
\usage{
is.grib(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
Returns a logical
}
\description{
\code{is.grib} tests whether and object is a GRIB object
}
\seealso{
\code{\link{is.grib.message}}
}

